% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{cdp_cox}
\alias{cdp_cox}
\title{CDP Cox regression}
\usage{
cdp_cox(
  obs,
  epsilon,
  delta = 0.001,
  niters = NA,
  C_beta = 1,
  C_z = 1,
  sensitivity = NA,
  stepsize = 0.5,
  cutoff = 1
)
}
\arguments{
\item{obs}{n*(2+d) dimensional matrix. Output from sim_observations().}

\item{epsilon}{Positive numeric. Epsilon privacy budget in terms of DP.}

\item{delta}{Optional positive numeric. Delta privacy budget in terms of DP.
Defaults to 0.001.}

\item{niters}{Optional positive integer. Number of gradient descent iterations.
Defaults to 6*log(C_z^2 *n/p^2).}

\item{C_beta}{Optional positive numeric. l_2 upper bound for coefficients. Defaults to 1.}

\item{C_z}{Optional positive numeric. l_2 upper bound for covariates. Defaults to 1.}

\item{sensitivity}{Optional positive numeric. l_2-sensitivity of the gradient of the partial
log-likelihood. Defaults to an upper bound for ensuring privacy based on the other parameter values.}

\item{stepsize}{Optional positive numeric. Step size in the descent direction. Defaults to 0.5.}
}
\value{
A d-dimensional numeric vector estimating the Cox regression coefficients.
}
\description{
A private SGD implementation to estimate Cox regression coefficients, that satisfies
central differential privacy.
}
