% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{fdp_breslow}
\alias{fdp_breslow}
\title{FDP Breslow}
\usage{
fdp_breslow(
  datasets,
  epsilon,
  beta_hat,
  p_hat,
  delta = NA,
  cutofff = 1,
  C_z = 1,
  weights = NA,
  tree_height = NA
)
}
\arguments{
\item{datasets}{List of S datasets, with (2+d) columns, as generated from sim_observations().}

\item{epsilon}{Positive numeric vector of length S. Epsilon privacy budget in terms of DP.}

\item{beta_hat}{Numeric vector of length d. Estimate of the Cox regression coefficients}

\item{p_hat}{Positive numeric. Estimate of P(Y(cutoff)=1).}

\item{delta}{Optional positive numeric vector of length S. Delta privacy budget
in terms of DP. Defaults to S entries of 0.001.}

\item{weights}{Optional positive numeric vector of length S. Weights used to aggregate
the per-server cumulative hazard functions. Defaults to effective sample sizes, normalised
to sum to 1.}

\item{tree_height}{Optional positive numeric. Number of tree levels to}

\item{cutoff}{Optional positive numeric. Defaults to 1.}
}
\value{
list containing two positive numeric vectors, 'times' and 'vals'.
vals\link{t} is the estimate for the cumulative hazard at times\link{t}.
}
\description{
A distributed variant of the Breslow estimator satisfying federated differential privacy.
}
