% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{fdp_cox}
\alias{fdp_cox}
\title{FDP Cox regression}
\usage{
fdp_cox(
  obs,
  epsilon,
  delta = NA,
  niters = NA,
  C_beta = 1,
  C_z = 1,
  sensitivity = NA,
  stepsize = 0.5,
  weights = NA
)
}
\arguments{
\item{epsilon}{Positive numeric vector of length S. Epsilon privacy budget in terms of DP.}

\item{delta}{Optional positive numeric vector of length S. Delta privacy budget
in terms of DP. Defaults to S entries of 0.001.}

\item{niters}{Optional positive integer. Number of gradient descent iterations.
Defaults to 6*log(C_z^2 *sum(samples)/p^2).}

\item{C_beta}{Optional positive numeric. l_2 upper bound for coefficients. Defaults to 1.}

\item{C_z}{Optional positive numeric. l_2 upper bound for covariates. Defaults to 1.}

\item{sensitivity}{Optional positive numeric vector of length S. l_2-sensitivity of
the gradient of the partial log-likelihood. Defaults to an upper bound for ensuring
privacy based on the other parameter values.}

\item{stepsize}{Optional positive numeric. Step size in the descent direction. Defaults to 0.5.}

\item{datasets}{List of S datasets, with (2+d) columns, as generated from sim_observations().}

\item{weights.}{Optional positive numeric vector of length S. Weights used to aggregate
the gradient estimates. Defaults to effective sample sizes, normalised to sum to 1.}
}
\value{
A d-dimensional numeric vector estimating the Cox regression coefficients.
}
\description{
A private SGD implementation to estimate Cox regression coefficients, that satisfies
federated differential privacy.
}
