% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{fdp_probabilities}
\alias{fdp_probabilities}
\title{FDP marginal survival estimation}
\usage{
fdp_probabilities(times, epsilon, delta = NA, cutoff = 1)
}
\arguments{
\item{times}{list of S non-negative numeric vectors.}

\item{epsilon}{Positive numeric vector of length S. Epsilon privacy budget in terms of DP.}

\item{delta}{Optional positive numeric vector of length S. Delta privacy budget
in terms of DP. Defaults to S entries of 0.001.}

\item{cutoff.}{Optional positive numeric. Timepoint at which to estimate probability
of being at-risk. Defaults to 1.}
}
\value{
A numeric that is an FDP estimate of P(Y(1)=cutoff).
}
\description{
Marginal survival estimation via mean estimation with federated differential
privacy guarantees.
}
