% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{sim_observations}
\alias{sim_observations}
\title{Generate a dataset}
\usage{
sim_observations(n, beta, censor = 1, C_z = 1)
}
\arguments{
\item{n}{Integer. Number of observations to simulate.}

\item{beta}{Numeric vector. d-dimensional vector of the Cox coefficients.}

\item{censor}{Optional positive numeric. Censoring distribution is Exp(censor). Defaults to 1.}

\item{C_z}{Optional positive numeric. Covariates for each observation are generated as
i.i.d. from Uniform(-C_z/sqrt(d), C_z/sqrt(d)). Defaults to 1.}
}
\value{
A n*(2+length(beta)) matrix, where the first column is the observed time,
the second column is the 1/0 (failure observed/censored) and remaining columns are covariates.
}
\description{
Generates observations, with i.i.d. covariates from a uniform distribution,
and a failure time from the Cox proportional hazards model with baseline hazard 1.
The recorded observation may be censored according to an exponential distribution.
}
